refactor: move pagination to onboardingPagination
