+++ app/components/index.ts
@@ -1,3 +1,4 @@
 export * from './FavoritesHeader/FavoritesHeader';
 export * from './BottomSheet/BottomSheet';
 export * from './OnboardingMask/OnboardingMask';
+export * from './OnboardingPagination/OnboardingPagination';
